requirements.txt
flask
langchain-groq
langchain
langgraph
sqlvalidator
cryptography
mysql-connector-python
sqlalchemy
python-dotenv